Refactored GET functionality into InternalGetService, removed SecretService, updated tests accordingly
